+++ internal/ui/flash/flash_test.go
@@ -2,10 +2,12 @@ package flash
 
 import (
 	"errors"
+	"fmt"
 	"strings"
 	"testing"
 
 	"github.com/idursun/jjui/internal/ui/common"
+	"github.com/idursun/jjui/internal/ui/intents"
 	"github.com/idursun/jjui/internal/ui/layout"
 	"github.com/idursun/jjui/internal/ui/render"
 	"github.com/stretchr/testify/assert"
@@ -98,5 +100,49 @@ func TestDeleteOldest_RemovesFirstMessage(t *testing.T) {
 
 	if assert.Len(t, m.messages, 1) {
 		assert.Equal(t, "second", m.messages[0].text)
+	}
+}
+
+func TestHistory_CompletionBeforeRunning_ReconcilesByCommandID(t *testing.T) {
+	m := New()
+
+	cmd := m.Update(common.CommandCompletedMsg{ID: 9, Output: "ok", Err: nil})
+	assert.Nil(t, cmd)
+	assert.Empty(t, m.commandHistorySnapshot())
+	assert.Equal(t, 0, m.LiveMessagesCount())
+
+	cmd = m.Update(common.CommandRunningMsg{ID: 9, Command: "jj git fetch"})
+	assert.NotNil(t, cmd)
+	snapshot := m.commandHistorySnapshot()
+	if assert.Len(t, snapshot, 1) {
+		assert.Equal(t, "jj git fetch", snapshot[0].Command)
+		assert.Equal(t, "ok", snapshot[0].Text)
+	}
+}
+
+func TestHistory_IsBoundedToConfiguredLimit(t *testing.T) {
+	m := New()
+
+	for i := 1; i <= HistoryLimit+5; i++ {
+		m.AddWithCommand(fmt.Sprintf("m-%d", i), fmt.Sprintf("jj cmd %d", i), nil)
+	}
+
+	snapshot := m.commandHistorySnapshot()
+	if assert.Len(t, snapshot, HistoryLimit) {
+		assert.Equal(t, "m-6", snapshot[0].Text)
+		assert.Equal(t, "m-55", snapshot[len(snapshot)-1].Text)
+	}
+}
+
+func TestHistory_OnlyContainsCommandMessages(t *testing.T) {
+	m := New()
+
+	m.Update(intents.AddMessage{Text: "user message"})
+	m.AddWithCommand("command output", "jj status", nil)
+
+	snapshot := m.commandHistorySnapshot()
+	if assert.Len(t, snapshot, 1) {
+		assert.Equal(t, "jj status", snapshot[0].Command)
+		assert.Equal(t, "command output", snapshot[0].Text)
 	}
 }
